Implement the game with computer as codemaker
Create 3 ruby files containing the main classes. The
decoding_board file implements the classes representing
the game board which are responsible for holding most of the
game state data. The players file implements the classes
representing the players which are responsible for doing
most of the games actions. Finally, the mastermind file
implements the Game class, which represents the game logic,
responsible for controlling the game flow.